Improvment - Addition of custom model via extra-action
I modified two places in the document: Around line 352 so that a two
news extra_actions elements are available : "responseClass"  as well as
"model" in order to allow the user to define custom model for his custom
endpoints.
The other modification is on line 563+ to add the custom model onto the
modellist.