Fixed commit according to comment + few extra fixes
On the doc (index.rst), I added the parameter "required" to the model so
that one could indicate if the model is required or not.

On mapping.py:
Line 140, 415 and 421 the "required" field is given by the ['nullable']
of tastypie and not ['blank']. I corrected a bug as well where the
parameter was not read properly from the TastyPie model.